Updated Put and Delete

+++ Controllers/PizzaController.cs
@@ -50,17 +50,35 @@ public class PizzaController : ControllerBase
     [HttpPut("{id}")]
     public IActionResult Update(int id, Pizza pizza)
     {
+        if (id != pizza.Id)
+        {
+            return BadRequest();
+        }
+        var existingPizza = PizzaService.Get(id);
+
+        if (existingPizza == null)
+        {
+            return NotFound();
+        }
+
         PizzaService.Update(pizza);
 
-        return CreatedAtAction(nameof(Get), new { id = pizza.Id }, pizza);
+        return NoContent();
     }
 
     [HttpDelete("{id}")]
     public IActionResult Delete(int id)
     {
+        var pizza = PizzaService.Get(id);
+
+        if (pizza is null)
+        {
+            return NotFound();
+        }
+
         PizzaService.Delete(id);
 
-        return Ok("Pizza Deleted Successfully");
+        return NoContent();
     }
 
 }
